Add GeoJSON encoding for places in view, map JS
Removed mapbox API, since geocoding is now persisted to our database.
For each place, the name and location are now encoded as a GeoJSON
Feature, containing the coordinates as a Point  They are then parsed on
the JS side and added to a feature layer. This is extensible for two
reasons: we might switch to client-side rendering, thus a formalised
representation is best; or, we might want to add more than one place,
in which case the loop is abstract and trivial.

One problem surfaced: we store places from the geocoder as lat, lon,
but PostGIS and GeoJSON like them to be lon, lat. I should add a flip
function before persisting and migrate existing ones.