Add lite JSON API for getting all places
We will use that to create the map for all places without slowing
down initial load. For single places, we bake-in the data on the
server-side view. For all places, however, we might want to get
more clever about loading in the future, so let's have it via API
from the get-go.